+++ src/handlers/playlist.rs
@@ -68,6 +68,7 @@ pub fn handler(key: Key, app: &mut App) {
         app.playlist_offset = 0;
         if let Some(selected_playlist) = playlists.items.get(selected_playlist_index.to_owned()) {
           let playlist_id = selected_playlist.id.id().to_string();
+          app.active_playlist_id = Some(playlist_id.clone());
           app.dispatch(IoEvent::GetPlaylistTracks(playlist_id, app.playlist_offset));
         }
       };